refactor: split into modules classes
