+++ RSLang/src/scss/color.scss
@@ -1,0 +1,4 @@
+$backgroundColor: #ebebeb;
+$colorText: #1b1b1d;
+$highlightImportant: #ffc001;
+$assideColor: #015288;
