+++ RSLang/src/scss/footer.scss
@@ -1,0 +1,66 @@
+@import './mixin.scss';
+@import './color.scss';
+.footer {
+    width: 100%;
+    color: #fff;
+    &__container {
+        @include stretch-width;
+        min-height: 100px;
+        height: 100px;
+    }
+    &__orange {
+        background: $highlightImportant;
+        height: 100%;
+        @include center;
+        flex-direction: column;
+        padding: 0px 20px;
+        a {
+            transition: 0.2s;
+            &:hover {
+                transform: scale(1.2);
+            }
+        }
+    }
+    &__black {
+        background: #000;
+        width: 100%;
+        height: 100%;
+        display: flex;
+        justify-items: center;
+    }
+    &__list {
+        @include stretch-width;
+        width: 100%;
+        padding: 0px 100px;
+        li {
+            a {
+                transition: 0.2s;
+                display: flex;
+                align-items: center;
+                &:hover {
+                    color: $highlightImportant;
+                    svg {
+                        path {
+                            fill: $highlightImportant;
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+@media screen and (max-width: 1000px) {
+    .footer {
+        &__list {
+            padding: 0 30px;
+            flex-wrap: wrap;
+        }
+    }
+}
+@media screen and (max-width: 670px) {
+    .footer {
+        &__container {
+            height: 130px;
+        }
+    }
+}
