+++ RSLang/src/scss/global.scss
@@ -1,13 +1,11 @@
-@mixin stretch-width {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-@mixin center {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
+@import './aside.scss';
+@import './color.scss';
+@import './footer.scss';
+@import './header.scss';
+@import './home.scss';
+@import './main.scss';
+@import './mixin.scss';
+
 * {
     box-sizing: border-box;
 }
@@ -29,7 +27,7 @@ button {
 }
 a {
     text-decoration: none;
-    color: #fff;
+    color: $backgroundColor;
     cursor: pointer;
 }
 ul,
@@ -48,8 +46,8 @@ button {
 body {
     font-family: sans-serif;
     font-weight: 400;
-    background: #ebebeb;
-    color: #1b1b1d;
+    background: $backgroundColor;
+    color: $colorText;
     padding: 0;
     margin: 0;
     position: relative;
@@ -70,189 +68,9 @@ body {
         display: none;
     }
 }
-.header {
-    padding: 30px 0;
-    width: 100%;
-    &__menu {
-        @include center;
-        flex-direction: column;
-        padding: 0 20px;
-    }
-    &__register {
-        position: absolute;
-        top: 100px;
-        transition: 0.2s;
-        cursor: pointer;
-        z-index: 3;
-        &:hover {
-            transform: scale(1.2);
-        }
-    }
-}
-.main {
-    width: 100%;
-    min-height: 100vh;
-    @include stretch-width;
-    flex-direction: column;
-}
-.footer {
-    width: 100%;
-    color: #fff;
-    &__container {
-        @include stretch-width;
-        min-height: 100px;
-        height: 100px;
-    }
-    &__orange {
-        background: #ffc001;
-        height: 100%;
-        @include center;
-        flex-direction: column;
-        padding: 0px 20px;
-        a {
-            transition: 0.2s;
-            &:hover {
-                transform: scale(1.2);
-            }
-        }
-    }
-    &__black {
-        background: #000;
-        width: 100%;
-        height: 100%;
-        display: flex;
-        justify-items: center;
-    }
-    &__list {
-        @include stretch-width;
-        width: 100%;
-        padding: 0px 100px;
-        li {
-            a {
-                transition: 0.2s;
-                display: flex;
-                align-items: center;
-                &:hover {
-                    color: #ffc001;
-                    svg {
-                        path {
-                            fill: #ffc001;
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-.work-pages {
-    height: 100%;
-    padding: 30px;
-    @include stretch-width();
-    align-items: start;
-    &__img {
-        max-height: 650px;
-        width: 100%;
-        overflow: hidden;
-    }
-}
-.start-page {
-    padding: 50px 0;
-    &__title {
-        font-size: 40px;
-        margin-bottom: 30px;
-        color: #ffc001;
-        max-width: 360px;
-        line-height: 35px;
-    }
-    &__text {
-        max-width: 550px;
-    }
-}
-.aside {
-    color: #fff;
-    background: #015288;
-    align-items: start;
-    flex-direction: column;
-    align-items: start;
-    padding: 30px;
-    &__el {
-        cursor: pointer;
-        margin: 10px;
-        transition: 0.2s;
-        &:hover {
-            color: #ffc001;
-        }
-    }
-    &__games {
-        padding-left: 20px;
-        li {
-            list-style-type: circle;
-        }
-    }
-}
-.team {
-    &__title {
-        margin-bottom: 30px;
-    }
-    &__list {
-        display: flex;
-        li {
-            display: flex;
-            flex-direction: column;
-            margin: 100px 40px 0 40px;
-            max-width: 150px;
-            text-align: center;
-            p {
-                font-size: 14px;
-            }
-            &:nth-child(2n + 1) {
-                margin-top: 0px;
-            }
-        }
-        a {
-            display: flex;
-            align-items: center;
-            color: black;
-        }
-        img {
-            max-width: 150px;
-            border-radius: 100px;
-        }
-    }
-}
-@media screen and (max-width: 1000px) {
-    .work-pages {
-        flex-direction: column;
-    }
-    .team {
-        &__list {
-            flex-wrap: wrap;
-            justify-content: space-around;
-            li {
-                margin-top: 0px;
-            }
-        }
-    }
-    .footer {
-        &__list {
-            padding: 0 30px;
-            flex-wrap: wrap;
-        }
-    }
-}
+
+
 @media screen and (max-width: 670px) {
-    .footer {
-        &__container {
-            height: 130px;
-        }
-    }
-    .aside {
-        display: none;
-        position: absolute;
-        height: 100vh;
-        padding-top: 50px;
-        z-index: 4;
-    }
     .root {
         &__burger {
             display: flex;
