+++ RSLang/src/scss/header.scss
@@ -1,0 +1,20 @@
+@import './mixin.scss';
+.header {
+    padding: 30px 0;
+    width: 100%;
+    &__menu {
+        @include center;
+        flex-direction: column;
+        padding: 0 20px;
+    }
+    &__register {
+        position: absolute;
+        top: 100px;
+        transition: 0.2s;
+        cursor: pointer;
+        z-index: 3;
+        &:hover {
+            transform: scale(1.2);
+        }
+    }
+}
