+++ RSLang/src/scss/mixin.scss
@@ -1,0 +1,10 @@
+@mixin stretch-width {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+@mixin center {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
