+++ RSLang/src/view/pages/book.ts
@@ -1,0 +1,8 @@
+class Book {
+  sectionBook(): string {
+    return `
+        <h3>Учебник</h3>
+        `;
+  }
+}
+export default Book;
