+++ RSLang/src/view/pages/game.ts
@@ -1,0 +1,8 @@
+class Game {
+  sectionGame(): string {
+    return `
+        <h3>Мини игры</h3>
+        `;
+  }
+}
+export default Game;
