+++ RSLang/src/view/pages/statistic.ts
@@ -1,0 +1,9 @@
+class Statistic {
+  sectionStatistic(): string {
+    return `
+        <h3>Статистика</h3>
+        <p>Доступно только для зарегестрированный пользователей</p>
+        `;
+  }
+}
+export default Statistic;
